+++ core/src/com/color/game/command/PushCommand.java
@@ -3,6 +3,10 @@ package com.color.game.command;
 import com.badlogic.gdx.math.Vector2;
 import com.color.game.elements.dynamicelements.BaseDynamicElement;
 
+/**
+ * Special command to push a BaseDynamicElement in a certain direction according to the force
+ * to exert on the element, this force will be exert until the command is ended
+ */
 public class PushCommand implements Command {
 
     private Vector2 force;